handel the protection of home page and profile + fix
