+++ Front_end/admin.html
@@ -37,7 +37,7 @@
             </nav>
         </header>
         <main>
-            
+            Coooc c'est admin
         </main>
 
     </body>
