--- Front_end/exemple.html
@@ -1,23 +1,0 @@
-
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <title> MOVIEW </title>
-        <link rel="stylesheet" href="Style/styleteste.css">
-        <!-- pour les icones -->
-        <link href='https://unpkg.com/boxicons@2.1.4/css/boxicons.min.css' rel='stylesheet'>
-    </head>
-    <body class="profil-body">
-        <header class="header">
-            <h2 class="logo">MOVIEW</h2> 
-            <nav class="navigation">
-            </nav>
-        </header>
-        <div class="parent">
-            <div class="child1">A</div>
-            <div class="child2">B</div>
-          </div>          
-    </body>
-    <script src="script.js"></script>
-</html>
