+++ Front_end/home.html
@@ -1,0 +1,42 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <title> MOVIEW </title>
+        <link rel="stylesheet" href="Style/style.css">
+        <!-- pour les icones -->
+        <link href='https://unpkg.com/boxicons@2.1.4/css/boxicons.min.css' rel='stylesheet'>
+    </head>
+    <body  onload="profileUser()"class="login-body">
+        <header>
+            <h2 class="logo">MOVIEW</h2> 
+            <nav class="navigation">
+                <ul class= "menu"id="menu">
+                  <li id="menu-login">
+                    <a href="profil.html" class="nav-link" id="username">
+                        <span class="link-text"></span>
+                    </a>
+                </li>
+                  <li id="menu-register">
+                    <a href="register.html" class="nav-link">
+                        <span class="link-text">REGISTER</span>
+                    </a>
+                </li>
+                  <li id="menu-films">
+                    <a href="films.html" class="nav-link">
+                        <span class="link-text">FILMS</span>
+                    </a>
+                </li>
+                  <li id="menu-lists">
+                    <a href="lists.html" class="nav-link">
+                        <span class="link-text">LISTS</span>
+                    </a>
+                </li>
+                </ul>
+            </nav>
+
+        </header>
+
+    </body>
+    <script src="script.js"></script>
+</html>
