--- Front_end/style/styleteste.css
@@ -1,33 +1,0 @@
-header{
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    padding: 2px 100px;
-    display: flex;
-    justify-content: space-between;
-    align-items: flex-start;
-    position: absolute;
-    background-color: green;
-}
-
-.parent {
-    top:auto;
-    background: lightgray;
-    padding: 10px;
-    position: absolute;
-    z-index: 5;
-    align-items: center;
-    top: 6em;
-    
-  }
-  
-  .child1 {
-    background: red;
-    margin-bottom: 10px;
-  }
-  
-  .child2 {
-    background: blue;
-  }
-
